Predicting-House-Prices/requirements.txt
lightgbm
pandas
Pillow
plotly
scikit_learn
streamlit
xgboost
